trying .long for data dtype

+++ gene_transformer/dataset.py
@@ -55,6 +55,6 @@ class FASTADataset(Dataset):  # type: ignore[type-arg]
         return len(self.sequences)
 
     def __getitem__(self, idx: int) -> torch.Tensor:
-        return self.sequences[idx]  # type:ignore[no-any-return]
+        return self.sequences[idx].long()  # type:ignore[no-any-return]
